Prevent problems with signing aggregated packages
Without this patch, signing of aggregated packages with signing keys newer than the package build time fails.

When OBS aggregates a package into another project,
it tries to do so using the original build time.
However, if the signing key in the new project is newer than the package build time,
this fails giving errors like this in /var/log/signer.log:

  ...
  using signtime 1505324598
  gpg: key has been created 23557508 seconds in future (time warp or clock problem)
  gpg: signing failed: Time conflict
  sign failed: 512 - checking digest
  ...

In the OBS frontend, the package will stay in the "finished" state.

Using the gpg --ignore-time-conflict option, this error is only treated as a warning.