Make it possible to configure multiple agent socket locations to try
Unfortunately gpg uses a different socket depending on the existance
of the /run/user/<uid> directory. This means that the socket location
is different if the user is logged in or not.

As calling gpgconf --list-dirs all the time is somewhat costly we
allow to configure the possible socket locations. If the socket
cannot be found or the agent socket list is empty, gpgconf --list-dirs
will be used as fallback.