Support on-disk aliases for user names
We did this with 'map' statements in the config file before. We
now can read aliases from disk, so that we do not need to restart
the signer everytime we want to change a mapping.

This works similar to the .phrases dir, except the file content
is a key name instead of a passphrase.