Add a fingerprint subpacket to v4 signatures
We do this via the pubalgo probe: we use the special 04040404 hash
to signal that we accept a v4 signature as return. We then get
the fingerprint from this signature and use it to construct the
signed data.

This does not seem to work with the old files-are-digests patch,
because it will always return a v3 signature even if --force-v3-sigs
is not used. (We could fix the patch, but it is probably not worth
the hassle.)