Implement signing with libgcrypt for unprotected privkeys
This is off by default, it can be enabled by setting 'use-gcrypt'
in the config.
The reason for this is that gpg-agent does not allow any
parallelism in its request (despite making use of threads).

This means that we run into a scalability problem when there
are many requests. This is even more the case for rsa@4096
pubkeys where signature generation is slow.